agriops_env/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agriops-env"
version = "1.0.0"
description = "AgriOpsEnv benchmark for OpenEnv hackathon"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Databaes" }]
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agriops_env*"]

[tool.uv]
dev-dependencies = ["pytest>=8.0", "httpx>=0.27"]

[project.scripts]
server = "server.app:main"

